@@ -5,12 +5,11 @@
     "prune-react-copies": "node ./scripts/prune-react-copies.js",
     "bootstrap": "lerna bootstrap && npm run prune-react-copies",
     "install-react": "rm -rf node_modules/react{,-dom,-addons-test-utils} && npm install",
-    "install-react:0.12": "npm run install-react -- react@0.12",
     "install-react:0.13": "npm run install-react -- react@0.13",
     "install-react:0.14": "npm run install-react -- react@0.14 react-dom@0.14 react-addons-test-utils@0.14",
     "install-react:15": "npm run install-react -- react@15 react-dom@15 react-addons-test-utils@15",
     "lint": "eslint --ext '.jsx,.js' ./",
-    "build-babel": "babel -D --ignore '__tests__,__mocks__' packages/$PACKAGE/src -d packages/$PACKAGE/lib",
+    "build-babel": "babel -D --ignore '__tests__,__mocks__,__coverage__' packages/$PACKAGE/src -d packages/$PACKAGE/lib",
     "build-webpack": "webpack --config packages/$PACKAGE/webpack.config.js",
     "build:component-playground": "PACKAGE=react-component-playground npm run build-webpack",
     "build:component-tree": "PACKAGE=react-component-tree npm run build-babel",
@@ -25,7 +24,8 @@
     "test-jest": "jest --coverage --config packages/$PACKAGE/jest.config.json",
     "test:cosmos": "PACKAGE=react-cosmos npm run test-jest",
     "test:cosmos-webpack": "PACKAGE=react-cosmos-webpack npm run test-jest",
-    "test": "npm run lint && npm run test-karma && npm run test:cosmos && npm run test:cosmos-webpack",
+    "test:redux-proxy": "PACKAGE=react-cosmos-redux-proxy npm run test-jest",
+    "test": "npm run lint && npm run test-karma && npm run test:cosmos && npm run test:cosmos-webpack && npm run test:redux-proxy",
     "coveralls": "lcov-result-merger '{packages/*/src/__coverage__,test/coverage/*}/lcov.info' | node_modules/coveralls/bin/coveralls.js"
   },
   "dependencies": {
@@ -41,6 +41,7 @@
     "chai": "^3.5.0",
     "coveralls": "^2.11.12",
     "css-loader": "^0.23.1",
+    "enzyme": "^2.4.1",
     "eslint": "^3.7.1",
     "eslint-config-airbnb": "^10.0.1",
     "eslint-plugin-import": "^1.16.0",
@@ -65,7 +66,7 @@
     "less": "^2.7.1",
     "less-loader": "^2.2.3",
     "mocha": "^3.0.2",
-    "react": ">=0.12 <16",
+    "react": ">=0.13 <16",
     "react-addons-test-utils": "<16",
     "react-dom": "<16",
     "rimraf": "^2.5.4",
