@@ -10,6 +10,6 @@
     "react-dom-polyfill": "^1.0.0-beta.4"
   },
   "peerDependencies": {
-    "react": ">=0.12 <16"
+    "react": ">=0.13 <16"
   }
 }
