@@ -6,10 +6,10 @@
   "license": "MIT",
   "main": "lib/index.js",
   "dependencies": {
-    "lodash": "^4.15.0",
+    "lodash.omit": "^4.5.0",
     "redux": "^3.6.0"
   },
   "peerDependencies": {
-    "react": ">=0.12 <16"
+    "react": ">=0.13 <16"
   }
 }
