@@ -1,6 +1,6 @@
 export const WELL_ROWS = 20;
 export const WELL_COLS = 10;
-export const DROP_FRAMES_DEFAULT = 48;
+export const DROP_FRAMES_DEFAULT = 48; // 1 row in less than a second
 export const DROP_FRAMES_DECREMENT = 1.5;
-export const DROP_FRAMES_ACCELERATED = 1.5;
+export const DROP_FRAMES_ACCELERATED = 1.5; // 40 rows per second
 export const LINE_CLEAR_BONUSES = [100, 300, 500, 800];
