@@ -1,12 +1,13 @@
 const FakeRouter = jest.fn();
 const FakeComponentPlayground = () => {};
-const fakeProxies = [];
+const fakeProxies = [{}, {}];
 const fakeComponentsInput = {};
 const fakeFixturesInput = {};
 const fakeComponentsOutput = {};
 const fakeFixturesOutput = {};
 const fakeLoadComponents = jest.fn(() => fakeComponentsOutput);
 const fakeLoadFixtures = jest.fn(() => fakeFixturesOutput);
+const fakeLinkedList = {};
 
 jest.mock('react-querystring-router', () => ({ Router: FakeRouter }));
 jest.mock('react-component-playground', () => FakeComponentPlayground);
@@ -14,7 +15,10 @@ jest.mock('../load-modules', () => ({
   loadComponents: fakeLoadComponents,
   loadFixtures: fakeLoadFixtures,
 }));
+jest.mock('../linked-list', () => jest.fn(() => fakeLinkedList));
 
+const PreviewLoader = require('../proxies/PreviewLoader').default;
+const createLinkedList = require('../linked-list');
 const startReactCosmos = require('../index');
 
 const routerInstance = startReactCosmos({
@@ -28,9 +32,18 @@ it('uses Componet Playground as router component class', () => {
   expect(getComponentClass()).toBe(FakeComponentPlayground);
 });
 
-it('sends proxies to Component Playground props', () => {
-  const { proxies } = FakeRouter.mock.calls[0][0].defaultProps;
-  expect(proxies).toBe(fakeProxies);
+it('create proxy list with user proxies', () => {
+  expect(createLinkedList.mock.calls[0][0][0]).toBe(fakeProxies[0]);
+  expect(createLinkedList.mock.calls[0][0][1]).toBe(fakeProxies[1]);
+});
+
+it('end proxy list with internal proxies', () => {
+  expect(createLinkedList.mock.calls[0][0][3]).toBe(PreviewLoader);
+});
+
+it('sends proxy list to Component Playground props', () => {
+  const { firstProxy } = FakeRouter.mock.calls[0][0].defaultProps;
+  expect(firstProxy).toBe(fakeLinkedList);
 });
 
 it('sends components input to loadComponents lib', () => {
