@@ -1,6 +1,9 @@
 import ReactQuerystringRouter from 'react-querystring-router';
 import ComponentPlayground from 'react-component-playground';
 import { loadComponents, loadFixtures } from './load-modules';
+import createLinkedList from './linked-list';
+import PreviewLoader from './proxies/PreviewLoader';
+import createStateProxy from './proxies/StateProxy';
 
 const getTitleForFixture = (params) => {
   let title = 'React Cosmos';
@@ -17,11 +20,19 @@ module.exports = ({
   proxies,
   components,
   fixtures,
-}) =>
-  new ReactQuerystringRouter.Router({
+}) => {
+  const firstProxy = createLinkedList([
+    ...proxies,
+    // Loaded by default in all configs
+    createStateProxy(),
+    // The final proxy in the chain simply renders the preview component
+    PreviewLoader,
+  ]);
+
+  return new ReactQuerystringRouter.Router({
     container: document.body.appendChild(document.createElement('div')),
     defaultProps: {
-      proxies,
+      firstProxy,
       components: loadComponents(components),
       fixtures: loadFixtures(fixtures),
     },
@@ -30,3 +41,4 @@ module.exports = ({
       document.title = getTitleForFixture(params);
     },
   });
+};
