@@ -4,8 +4,9 @@ module.exports = (React) => {
   if (version >= 0.14) {
     // Let bundlers (e.g. webpack) know react-dom won't always be there
     try {
-      // eslint-disable-next-line global-require, import/no-extraneous-dependencies
+      /* eslint-disable global-require, import/no-extraneous-dependencies, import/no-unresolved */
       return require('react-dom');
+      /* eslint-enable global-require, import/no-extraneous-dependencies, import/no-unresolved */
     } catch (e) {
       return null;
     }
